+++ services/kube-eleven/server/adapters/inbound/grpc/adapter.go
@@ -1,0 +1,52 @@
+package grpc
+
+import (
+	"fmt"
+	"net"
+
+	"github.com/rs/zerolog/log"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/health"
+	"google.golang.org/grpc/health/grpc_health_v1"
+
+	"github.com/berops/claudie/internal/utils"
+	"github.com/berops/claudie/proto/pb"
+)
+
+const (
+	defaultPort = 50054
+)
+
+type GrpcAdapter struct {
+	tcpListener       net.Listener
+	server            *grpc.Server
+	healthcheckServer *health.Server
+}
+
+func CreateGrpcAdapter() *GrpcAdapter {
+	var (
+		g   = &GrpcAdapter{}
+		err error
+	)
+
+	port := utils.GetenvOr("KUBE_ELEVEN_PORT", fmt.Sprint(defaultPort))
+	bindingAddress := net.JoinHostPort("0.0.0.0", port)
+	g.tcpListener, err = net.Listen("tcp", bindingAddress)
+	if err != nil {
+		log.Fatal().Msgf("Failed to bind to %s : %v", bindingAddress, err)
+	}
+	log.Info().Msgf("Kube-eleven microservice is listening on %s", bindingAddress)
+
+	g.server = grpc.NewServer()
+	pb.RegisterKubeElevenServiceServer(g.server, &KubeElevenGrpcService{})
+
+	// Add healthcheck service to the gRPC server
+	g.healthcheckServer = health.NewServer()
+	// Kube-eleven does not have any custom healthcheck functions,
+	// thus always serving.
+	g.healthcheckServer.SetServingStatus("kube-eleven-liveness", grpc_health_v1.HealthCheckResponse_SERVING)
+	g.healthcheckServer.SetServingStatus("kube-eleven-readiness", grpc_health_v1.HealthCheckResponse_SERVING)
+	grpc_health_v1.RegisterHealthServer(g.server, g.healthcheckServer)
+
+	return g
+}
