+++ services/kube-eleven/server/main.go
@@ -1,0 +1,66 @@
+package main
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"os"
+	"os/signal"
+	"syscall"
+	"time"
+
+	"github.com/rs/zerolog/log"
+	"golang.org/x/sync/errgroup"
+
+	"github.com/berops/claudie/internal/utils"
+)
+
+func main() {
+	// Initialize logger
+	utils.InitLog("kube-eleven")
+
+	g, ctx := errgroup.WithContext(context.Background())
+
+	//goroutine for interrupt
+	g.Go(func() error {
+		ch := make(chan os.Signal, 1)
+		signal.Notify(ch, os.Interrupt, syscall.SIGTERM)
+		defer signal.Stop(ch)
+
+		// wait for either the received signal or
+		// check if an error occurred in other
+		// go-routines.
+		var err error
+		select {
+		case <-ctx.Done():
+			err = ctx.Err()
+		case sig := <-ch:
+			log.Info().Msgf("Received signal %v", sig)
+			err = errors.New("interrupt signal")
+		}
+
+		log.Info().Msg("Gracefully shutting down gRPC server")
+		s.GracefulStop()
+		healthServer.Shutdown()
+
+		// Sometimes when the container terminates gRPC logs the following message:
+		// rpc error: code = Unknown desc = Error: No such container: hash of the container...
+		// It does not affect anything as everything will get terminated gracefully
+		// this time.Sleep fixes it so that the message won't be logged.
+		time.Sleep(1 * time.Second)
+
+		return err
+	})
+
+	//server goroutine
+	g.Go(func() error {
+		// s.Serve() will create a service goroutine for each connection
+		if err := s.Serve(lis); err != nil {
+			return fmt.Errorf("kube-eleven failed to serve: %w", err)
+		}
+		log.Info().Msg("Finished listening for incoming connections")
+		return nil
+	})
+
+	log.Info().Msgf("Stopping Kube-eleven: %s", g.Wait())
+}
